Move shutdown script out of commitments
